fix(db): ajustar tipos de datos, SQL nativo de NPCs y serializacion de tags

+++ backend/alembic/versions/001_initial_migration.py
@@ -24,7 +24,7 @@ def upgrade() -> None:
     # Crear tabla zonas
     op.create_table(
         "zonas",
-        sa.Column("id", sa.Integer(), primary_key=True),
+        sa.Column("id", sa.BigInteger(), primary_key=True),
         sa.Column("name", sa.String(256), nullable=False),
         sa.Column(
             "tags",
